chore: move highlight to prevent bundling
